Add main parser / example parser
