--- mlbench/refimpls/pytorch/configs/debug-cifar10.json
@@ -1,20 +1,0 @@
-{
-  "meta": {
-    "debug": true,
-    "save": "best"
-  },
-  "controlflow": {
-    "num_epochs": 3
-  },
-  "dataset": {
-    "name": "cifar10",
-    "batch_size": 128,
-    "num_workers": 2,
-    "reshuffle_per_epoch": true,
-    "num_classes": 10
-  },
-  "model": {
-    "name": "resnet18_bkj"
-  },
-  "runtime": {}
-}
